fixes and changes
Yeah just some fixes i did as well as removing some code that got repeated a bunch of times. I still have a bug for when you open the map that you get it 3 times instead of 2 times but i will surely fix that later on. Not sure if i changed something in the Map.py file but i just put it in here just in case