Time Feature added
This update includes a Time system that display the time. Currently is a alone standing system and will later be integrated into the main loop. Every month is 30 days long  for simplicity reasons. Also the first row of buttons currently do nothing will be fixed later on. As those buttons are probably only for fighting or so later on.