Completed Time System
with this the Time System is fully integrated into the normal game loop and works perfectly fine. 
The first row of Times is still not functional as those are mainly for combat and for that i will need to rewrite the whole combat system which i will probably do far later on.