move with WASD, Time loop fix
You can travel on the map with WASD with is really amazing and makes it far easier then before. I also updated the time loop and made some changes to it as well as to the cultivation loop so the code runs way faster now and you can spam click the 50 years with no delay. It is still not perfect but for now that is good enough.