Time Manager
I added a Time Manager system that appears in an overlay where you will be able to choose your schedule for every hour still a work in progress i just wanted to update as it is the last day of the year. But surely i will get this to work next year.